Organize pod2 submodules
