+++ pod2/src/pod/circuit/mod.rs
@@ -4,4 +4,4 @@ pub mod operation;
 pub mod origin;
 pub mod pod;
 pub mod statement;
-mod util;
+pub mod util;
